Move TrustManager test to own file in Xamarin.Android.Net directory
